Split PAPR mask file into 5 files
Before this (and splitting up the notes), it was over 1500 lines and was making Dragon very slow. Now each file is < 400 lines, which should make it much more manageable.

Other than splitting the files, the only change I intended to include in this commit was to move the eye-lasers earlier (to setup.py).

Unfortunately, I forgot to make this commit, and then edited intake.py a bunch. So this commit now also includes some work on the new intake design.